feat: Consolidate Katana DAOs and scope reward strategies per org
