+++ src/actions/capitalDistributor/constants/capitalDistributorDaoSlotId.ts
@@ -1,0 +1,3 @@
+export enum CapitalDistributorDaoSlotId {
+    CAPITAL_DISTRIBUTOR_MEMBERS_FILE_DOWNLOAD = 'CAPITAL_DISTRIBUTOR_MEMBERS_FILE_DOWNLOAD',
+}
