+++ src/daos/cryptex/rewards/components/cryptexRewardsMembersFileDownload/index.ts
@@ -1,0 +1,2 @@
+export { CryptexRewardsMembersFileDownload } from './cryptexRewardsMembersFileDownload';
+export type { ICryptexRewardsMembersFileDownloadProps } from './cryptexRewardsMembersFileDownload.api';
