+++ src/daos/cryptex/rewards/constants/cryptexRewardsDialogId.ts
@@ -1,0 +1,3 @@
+export enum CryptexRewardsDialogId {
+    CRYPTEX_REWARDS_MEMBERS_FILE_DOWNLOAD = 'CRYPTEX_REWARDS_MEMBERS_FILE_DOWNLOAD',
+}
