+++ src/daos/cryptex/rewards/dialogs/cryptexRewardsMembersFileDownloadDialog/index.ts
@@ -1,0 +1,12 @@
+import dynamic from 'next/dynamic';
+
+export const CryptexRewardsMembersFileDownloadDialog = dynamic(() =>
+    import('./cryptexRewardsMembersFileDownloadDialog').then(
+        (mod) => mod.CryptexRewardsMembersFileDownloadDialog,
+    ),
+);
+
+export type {
+    ICryptexRewardsMembersFileDownloadDialogParams,
+    ICryptexRewardsMembersFileDownloadDialogProps,
+} from './cryptexRewardsMembersFileDownloadDialog';
