+++ src/daos/index.ts
@@ -1,17 +1,15 @@
 import { initialiseAragonDemo } from './aragonDemo';
 import { initialiseAragonDemoPolicies } from './aragonDemoPolicies';
 import { initialiseBoundless } from './boundless';
-import { initialiseCapitalDistributorTest } from './capitalDistributorTest';
 import { initialiseCryptex } from './cryptex';
-import { initialiseKatanaCDDemo } from './katanaCDDemo';
-import { initialiseKatanaEmissionsTest } from './katanaEmissionsTest';
-import { gaugeRewardDialogsDefinitions } from './strategies/gaugeReward';
-import { tokenRewardDialogsDefinitions } from './strategies/tokenReward';
+import { cryptexRewardsDialogsDefinitions } from './cryptex/rewards';
+import { initialiseKatana } from './katana';
+import { katanaRewardsDialogsDefinitions } from './katana/rewards';
 import { initialiseXmaquina } from './xmaquina';
 
 export const capitalDistributorDialogsDefinitions = {
-    ...gaugeRewardDialogsDefinitions,
-    ...tokenRewardDialogsDefinitions,
+    ...katanaRewardsDialogsDefinitions,
+    ...cryptexRewardsDialogsDefinitions,
 };
 
 export const initialiseDaos = () => {
@@ -20,7 +18,5 @@ export const initialiseDaos = () => {
     initialiseCryptex();
     initialiseXmaquina();
     initialiseAragonDemoPolicies();
-    initialiseCapitalDistributorTest();
-    initialiseKatanaCDDemo();
-    initialiseKatanaEmissionsTest();
+    initialiseKatana();
 };
