+++ src/daos/katana/rewards/components/katanaRewardsMembersFileDownload/index.ts
@@ -1,0 +1,2 @@
+export { KatanaRewardsMembersFileDownload } from './katanaRewardsMembersFileDownload';
+export type { IKatanaRewardsMembersFileDownloadProps } from './katanaRewardsMembersFileDownload.api';
