+++ src/daos/katana/rewards/components/katanaRewardsMembersFileDownload/katanaRewardsMembersFileDownload.api.ts
@@ -1,0 +1,13 @@
+import type { IAsset } from '@/modules/finance/api/financeService';
+import type { IDao } from '@/shared/api/daoService';
+
+export interface IKatanaRewardsMembersFileDownloadProps {
+    /**
+     * DAO to display in the header.
+     */
+    dao: IDao;
+    /**
+     * Asset selected in the campaign creation form, passed via PluginSingleComponent.
+     */
+    asset?: IAsset;
+}
