+++ src/daos/katana/rewards/constants/katanaRewardsDialogId.ts
@@ -1,0 +1,3 @@
+export enum KatanaRewardsDialogId {
+    KATANA_REWARDS_MEMBERS_FILE_DOWNLOAD = 'KATANA_REWARDS_MEMBERS_FILE_DOWNLOAD',
+}
