+++ src/daos/katana/rewards/constants/katanaRewardsMinEpochId.ts
@@ -1,0 +1,5 @@
+/**
+ * Minimum valid epoch ID for Katana reward distributions.
+ * Epochs below this value are invalid because the system was not deployed yet.
+ */
+export const katanaRewardsMinEpochId = 1460;
