+++ src/daos/katana/rewards/dialogs/katanaRewardsMembersFileDownloadDialog/index.ts
@@ -1,0 +1,9 @@
+import dynamic from 'next/dynamic';
+
+export const KatanaRewardsMembersFileDownloadDialog = dynamic(() =>
+    import('./katanaRewardsMembersFileDownloadDialog').then(
+        (mod) => mod.KatanaRewardsMembersFileDownloadDialog,
+    ),
+);
+
+export type { IKatanaRewardsMembersFileDownloadDialogParams } from './katanaRewardsMembersFileDownloadDialog';
