+++ src/daos/katana/rewards/dialogs/katanaRewardsMembersFileDownloadDialog/katanaRewardsMembersFileDownloadDialog.tsx
@@ -1,0 +1,273 @@
+'use client';
+
+import {
+    Dialog,
+    DropdownContainer,
+    DropdownItem,
+    InputContainer,
+    InputNumber,
+    invariant,
+} from '@aragon/gov-ui-kit';
+import { useEffect, useMemo, useState } from 'react';
+import { type Hex, parseUnits } from 'viem';
+import type { IAsset } from '@/modules/finance/api/financeService';
+import {
+    useEpochMetrics,
+    useRewardDistribution,
+} from '@/plugins/gaugeVoterPlugin/api/gaugeVoterService';
+import type { IGaugeVoterPlugin } from '@/plugins/gaugeVoterPlugin/types';
+import type { AragonBackendServiceError } from '@/shared/api/aragonBackendService';
+import type { Network } from '@/shared/api/daoService';
+import type { IDialogComponentProps } from '@/shared/components/dialogProvider';
+import { useDialogContext } from '@/shared/components/dialogProvider';
+import { useTranslations } from '@/shared/components/translationsProvider';
+import { KatanaRewardsDialogId } from '../../constants/katanaRewardsDialogId';
+import { katanaRewardsMinEpochId } from '../../constants/katanaRewardsMinEpochId';
+import { rewardUtils } from '../../utils/rewardUtils';
+
+export interface IKatanaRewardsMembersFileDownloadDialogParams {
+    /**
+     * Gauge plugin.
+     */
+    gaugePlugin: IGaugeVoterPlugin;
+    /**
+     * Network of the DAO.
+     */
+    network: Network;
+    /**
+     * Asset selected in the campaign creation form.
+     */
+    asset?: IAsset;
+    /**
+     * Called with the downloaded file name after a successful download.
+     */
+    onDownload?: (fileName: string) => void;
+}
+
+export interface IKatanaRewardsMembersFileDownloadDialogProps
+    extends IDialogComponentProps<IKatanaRewardsMembersFileDownloadDialogParams> {}
+
+export const KatanaRewardsMembersFileDownloadDialog: React.FC<
+    IKatanaRewardsMembersFileDownloadDialogProps
+> = (props) => {
+    const { location } = props;
+
+    invariant(
+        location.params != null,
+        'KatanaRewardsMembersFileDownloadDialog: params must be defined',
+    );
+
+    const { gaugePlugin, network, asset } = location.params;
+
+    const { t } = useTranslations();
+    const { close } = useDialogContext();
+
+    const [totalAmount, setTotalAmount] = useState('');
+    const [epochId, setEpochId] = useState('');
+
+    const epochMetrics = useEpochMetrics({
+        urlParams: {
+            pluginAddress: gaugePlugin.address as Hex,
+            network,
+        },
+        queryParams: {},
+    });
+
+    const currentEpochId = epochMetrics.data?.epochId
+        ? Number(epochMetrics.data.epochId)
+        : undefined;
+
+    useEffect(() => {
+        if (currentEpochId != null && epochId === '') {
+            setEpochId(String(currentEpochId));
+        }
+    }, [currentEpochId, epochId]);
+
+    const validEpochs = useMemo(() => {
+        if (currentEpochId == null) {
+            return [];
+        }
+
+        const epochs: number[] = [];
+
+        const firstEpoch =
+            katanaRewardsMinEpochId > currentEpochId
+                ? 1
+                : katanaRewardsMinEpochId;
+
+        for (
+            let i = currentEpochId;
+            i >= firstEpoch && epochs.length < 100;
+            i--
+        ) {
+            epochs.push(i);
+        }
+
+        return epochs;
+    }, [currentEpochId]);
+
+    const tokenDecimals =
+        asset?.token.decimals ?? gaugePlugin.settings.token.decimals;
+    const tokenSymbol = asset?.token.symbol;
+
+    const totalAmountInUnits = useMemo(
+        () =>
+            totalAmount
+                ? parseUnits(totalAmount, tokenDecimals).toString()
+                : '',
+        [totalAmount, tokenDecimals],
+    );
+
+    const rewardDistribution = useRewardDistribution(
+        {
+            urlParams: {
+                pluginAddress: gaugePlugin.address as Hex,
+                network,
+                epochId: Number(epochId),
+            },
+            queryParams: {
+                rewardTotalAmount: totalAmountInUnits,
+            },
+        },
+        { enabled: false },
+    );
+
+    const errorCode =
+        rewardDistribution.error &&
+        (rewardDistribution.error as AragonBackendServiceError).code;
+    const errorKey =
+        errorCode &&
+        [
+            'epochWindowInvalid',
+            'epochVotingNotClosed',
+            'epochWindowExpired',
+            'epochNoActiveVoters',
+        ].includes(errorCode)
+            ? errorCode
+            : 'default';
+
+    const isFormValid = totalAmount !== '' && epochId !== '';
+
+    const handleGenerate = () => {
+        rewardDistribution.refetch().then((result) => {
+            if (result.data == null) {
+                return;
+            }
+
+            const rewardJson = rewardUtils.toRewardJson({
+                owners: result.data.owners,
+            });
+
+            const blob = new Blob([JSON.stringify(rewardJson, null, 2)], {
+                type: 'application/json',
+            });
+            const url = URL.createObjectURL(blob);
+            const anchor = document.createElement('a');
+            const fileName = `reward-distribution-epoch-${result.data.epoch}.json`;
+            anchor.href = url;
+            anchor.download = fileName;
+            anchor.click();
+            URL.revokeObjectURL(url);
+            location.params?.onDownload?.(fileName);
+            handleClose();
+        });
+    };
+
+    const handleClose = () => {
+        close(KatanaRewardsDialogId.KATANA_REWARDS_MEMBERS_FILE_DOWNLOAD);
+    };
+
+    const selectedEpochLabel =
+        epochId !== ''
+            ? t(
+                  'app.daos.capitalDistributorTest.capitalDistributorTestMembersFileDownloadDialog.epochOption',
+                  { epochId },
+              )
+            : t(
+                  'app.daos.capitalDistributorTest.capitalDistributorTestMembersFileDownloadDialog.epochSelectPlaceholder',
+              );
+
+    return (
+        <>
+            <Dialog.Header
+                description={t(
+                    'app.daos.capitalDistributorTest.capitalDistributorTestMembersFileDownloadDialog.description',
+                )}
+                title={t(
+                    'app.daos.capitalDistributorTest.capitalDistributorTestMembersFileDownloadDialog.title',
+                )}
+            />
+            <Dialog.Content>
+                <div className="flex flex-col gap-6 py-2">
+                    <InputContainer
+                        helpText={t(
+                            'app.daos.capitalDistributorTest.capitalDistributorTestMembersFileDownloadDialog.epochIdHelpText',
+                        )}
+                        id="epoch-select"
+                        label={t(
+                            'app.daos.capitalDistributorTest.capitalDistributorTestMembersFileDownloadDialog.epochIdLabel',
+                        )}
+                        useCustomWrapper={true}
+                    >
+                        <div className="w-fit">
+                            <DropdownContainer
+                                disabled={validEpochs.length === 0}
+                                label={selectedEpochLabel}
+                                size="md"
+                                variant="tertiary"
+                            >
+                                {validEpochs.map((epoch) => (
+                                    <DropdownItem
+                                        key={epoch}
+                                        onSelect={() =>
+                                            setEpochId(String(epoch))
+                                        }
+                                        selected={epochId === String(epoch)}
+                                    >
+                                        {t(
+                                            'app.daos.capitalDistributorTest.capitalDistributorTestMembersFileDownloadDialog.epochOption',
+                                            { epochId: epoch },
+                                        )}
+                                    </DropdownItem>
+                                ))}
+                            </DropdownContainer>
+                        </div>
+                    </InputContainer>
+                    <InputNumber
+                        label={t(
+                            'app.daos.capitalDistributorTest.capitalDistributorTestMembersFileDownloadDialog.amountLabel',
+                        )}
+                        min={0}
+                        onChange={setTotalAmount}
+                        placeholder="0"
+                        suffix={tokenSymbol}
+                        value={totalAmount}
+                    />
+                    {rewardDistribution.isError && (
+                        <p className="text-critical-600 text-sm">
+                            {t(
+                                `app.daos.capitalDistributorTest.capitalDistributorTestMembersFileDownloadDialog.error.${errorKey}`,
+                            )}
+                        </p>
+                    )}
+                </div>
+            </Dialog.Content>
+            <Dialog.Footer
+                primaryAction={{
+                    label: t(
+                        'app.daos.capitalDistributorTest.capitalDistributorTestMembersFileDownloadDialog.generate',
+                    ),
+                    onClick: handleGenerate,
+                    disabled: !isFormValid,
+                    isLoading: rewardDistribution.isFetching,
+                }}
+                secondaryAction={{
+                    label: t(
+                        'app.daos.capitalDistributorTest.capitalDistributorTestMembersFileDownloadDialog.cancel',
+                    ),
+                    onClick: handleClose,
+                }}
+            />
+        </>
+    );
+};
