+++ src/daos/katana/rewards/index.ts
@@ -1,0 +1,2 @@
+export { KatanaRewardsMembersFileDownload } from './components/katanaRewardsMembersFileDownload';
+export { katanaRewardsDialogsDefinitions } from './constants/katanaRewardsDialogsDefinitions';
