--- src/daos/strategies/gaugeReward/components/gaugeRewardMembersFileDownload/index.ts
@@ -1,2 +1,0 @@
-export { GaugeRewardMembersFileDownload } from './gaugeRewardMembersFileDownload';
-export type { IGaugeRewardMembersFileDownloadProps } from './gaugeRewardMembersFileDownload.api';
