--- src/daos/strategies/gaugeReward/constants/gaugeRewardDialogId.ts
@@ -1,3 +1,0 @@
-export enum GaugeRewardDialogId {
-    GAUGE_REWARD_MEMBERS_FILE_DOWNLOAD = 'GAUGE_REWARD_MEMBERS_FILE_DOWNLOAD',
-}
