--- src/daos/strategies/gaugeReward/constants/gaugeRewardMinEpochId.ts
@@ -1,5 +1,0 @@
-/**
- * Minimum valid epoch ID for gauge reward distributions.
- * Epochs below this value are invalid because the system was not deployed yet.
- */
-export const gaugeRewardMinEpochId = 1460;
