--- src/daos/strategies/gaugeReward/dialogs/gaugeRewardMembersFileDownloadDialog/index.ts
@@ -1,9 +1,0 @@
-import dynamic from 'next/dynamic';
-
-export const GaugeRewardMembersFileDownloadDialog = dynamic(() =>
-    import('./gaugeRewardMembersFileDownloadDialog').then(
-        (mod) => mod.GaugeRewardMembersFileDownloadDialog,
-    ),
-);
-
-export type { IGaugeRewardMembersFileDownloadDialogParams } from './gaugeRewardMembersFileDownloadDialog';
