--- src/daos/strategies/gaugeReward/index.ts
@@ -1,2 +1,0 @@
-export { GaugeRewardMembersFileDownload } from './components/gaugeRewardMembersFileDownload';
-export { gaugeRewardDialogsDefinitions } from './constants/gaugeRewardDialogsDefinitions';
