--- src/daos/strategies/tokenReward/components/tokenRewardMembersFileDownload/index.ts
@@ -1,2 +1,0 @@
-export { TokenRewardMembersFileDownload } from './tokenRewardMembersFileDownload';
-export type { ITokenRewardMembersFileDownloadProps } from './tokenRewardMembersFileDownload.api';
