--- src/daos/strategies/tokenReward/constants/tokenRewardDialogId.ts
@@ -1,3 +1,0 @@
-export enum TokenRewardDialogId {
-    TOKEN_REWARD_MEMBERS_FILE_DOWNLOAD = 'TOKEN_REWARD_MEMBERS_FILE_DOWNLOAD',
-}
