--- src/daos/strategies/tokenReward/dialogs/tokenRewardMembersFileDownloadDialog/index.ts
@@ -1,12 +1,0 @@
-import dynamic from 'next/dynamic';
-
-export const TokenRewardMembersFileDownloadDialog = dynamic(() =>
-    import('./tokenRewardMembersFileDownloadDialog').then(
-        (mod) => mod.TokenRewardMembersFileDownloadDialog,
-    ),
-);
-
-export type {
-    ITokenRewardMembersFileDownloadDialogParams,
-    ITokenRewardMembersFileDownloadDialogProps,
-} from './tokenRewardMembersFileDownloadDialog';
