--- src/daos/strategies/tokenReward/index.ts
@@ -1,2 +1,0 @@
-export { TokenRewardMembersFileDownload } from './components/tokenRewardMembersFileDownload';
-export { tokenRewardDialogsDefinitions } from './constants/tokenRewardDialogsDefinitions';
