--- src/modules/capitalFlow/constants/moduleDaoSlots.ts
@@ -1,3 +1,0 @@
-export enum CapitalFlowDaoSlotId {
-    CAPITAL_DISTRIBUTOR_MEMBERS_FILE_DOWNLOAD = 'CAPITAL_DISTRIBUTOR_MEMBERS_FILE_DOWNLOAD',
-}
